build(x3d-serializer): move code base to src directory for proper packaging
