+++ packages/io/x3d-serializer/package.json
@@ -4,7 +4,7 @@
   "description": "X3D Serializer for JSCAD",
   "homepage": "https://openjscad.xyz/",
   "repository": "https://github.com/jscad/OpenJSCAD.org",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "coverage": "nyc --all --reporter=html --reporter=text npm test",
     "test": "ava --verbose --timeout 2m './tests/*.test.js'"
