+++ packages/io/x3d-serializer/src/index.js
